tests/wpt/web-platform-tests/clipboard-apis/permissions/readText-denied.https.html

<!doctype html>
<meta charset="utf-8">
<title>navigator.clipboard.readText() fails when permission denied</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'denied');
  await promise_rejects_dom(t,
      'NotAllowedError', navigator.clipboard.readText());
}, 'navigator.clipboard.readText() fails when permission denied');
</script>
